\name{startmrca-package}
\alias{startmrca-package}
\alias{startmrca}
\docType{package}
\title{
\packageTitle{startmrca} startmrca
}
\description{
\packageDescription{startmrca} Estimates time to the common ancestor for a beneficial allele. 
}

\details{
\code{run.startmrca} is the only function in the package. See \code{help(run.startmrca)} for details.
}

\author{
\packageAuthor{startmrca} Joel Smith

Maintainer: \packageMaintainer{startmrca} Joel Smith
}
\references{
Joel Smith, Graham Coop, Matthew Stephens and John Novembre (2018): Estimating Time to the Common Ancestor for a Beneficial Allele. Molecular Biology and Evolution. https://doi.org/10.1093/molbev/msy006
}

\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
help(run.startmrca)
}
