\name{startmrca-package}
\alias{startmrca-package}
\alias{startmrca}
\docType{package}
\title{
\packageTitle{startmrca} startmrca
}
\description{
\packageDescription{startmrca} Estimates time to the common ancestor for a beneficial allele. 
}

\details{
run.startmrca is the only function in the package. See help(run.startmrca) for details
}

\author{
\packageAuthor{startmrca} Joel Smith

Maintainer: \packageMaintainer{startmrca} Joel Smith
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
See help(run.startmrca)
}
